+++ src/main/resources/database/init.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS "Currency"
     "id"        BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     "name"              VARCHAR(255) NOT NULL UNIQUE,
     "tag"               VARCHAR(10) NOT NULL UNIQUE,
-    "rate"              FLOAT NOT NULL,
+    "rate"              NUMERIC NOT NULL,
     "createdAt"         TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS "Ride"
     "id"        BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     "categoryId"    BIGINT REFERENCES "Category"("id"),
     "currencyId"    BIGINT NOT NULL REFERENCES "Currency"("id"),
-    "amount"        FLOAT NOT NULL,
+    "amount"        NUMERIC NOT NULL,
     "distance"      FLOAT NOT NULL,
     "passengers"    INTEGER NOT NULL,
     "tripType"      VARCHAR(255) NOT NULL,
